requirements.txt
streamlit==1.29.0
langgraph==0.0.62
langchain==0.1.0
langchain-core==0.1.10
langchain-google-genai==1.0.1
langchain-groq==0.1.3
python-dotenv==1.0.0